requirements.txt
asgiref==3.4.1
Django==3.2.16
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==8.4.0
pytz==2022.4
sqlparse==0.4.3
typing-extensions==4.1.1
whitenoise==5.3.0
